+++ packages/generator/src/oot/actors/Player.c
@@ -13,6 +13,7 @@
 #include <combo/entrance.h>
 #include <actors/En_Kanban/En_Kanban.h>
 #include <combo/oot/player_action.h>
+#include <actors/Custom_En_Bom/Custom_En_Bom.h>
 
 void ArrowCycle_Handle(Player* link, PlayState* play);
 void Ocarina_HandleCustomSongs(Player* link, PlayState* play);
@@ -405,16 +406,14 @@ static u16 blastMaskDelay(void)
 
 static void Player_BlastMask(PlayState* play, Player* link)
 {
-    Actor* bomb;
-    s16* bombTimer;
+    EnBomMM* bomb;
 
     if (gBlastMaskDelayAcc)
         return;
-    bomb = Actor_Spawn(&play->actorCtx, play, ACTOR_EN_BOM, link->actor.focus.pos.x, link->actor.focus.pos.y, link->actor.focus.pos.z, 0, 0, 0, 0);
+    bomb = (EnBomMM*)Actor_Spawn(&play->actorCtx, play, ACTOR_CUSTOM_EN_BOM, link->actor.focus.pos.x, link->actor.focus.pos.y, link->actor.focus.pos.z, BOMB_EXPLOSIVE_TYPE_BLAST, 0, 0, 0);
     if (!bomb)
         return;
-    bombTimer = (void*)((char*)bomb + 0x1e8);
-    *bombTimer = 2;
+    bomb->timer = 0;
     gBlastMaskDelayAcc = blastMaskDelay();
     Interface_LoadItemIconImpl(play, 0);
 }
@@ -478,6 +477,7 @@ static u8 sCustomActionModelGroups[] = {
     0xe, /* PLAYER_MODELGROUP_BOTTLE, PLAYER_CUSTOM_IA_SPRING_WATER */
     0xe, /* PLAYER_MODELGROUP_BOTTLE, PLAYER_CUSTOM_IA_SPRING_WATER_HOT */
     0xe, /* PLAYER_MODELGROUP_BOTTLE, PLAYER_CUSTOM_IA_ZORA_EGG */
+    0x7, /* PLAYER_MODELGROUP_EXPLOSIVES, PLAYER_CUSTOM_IA_POWDER_KEG */
 };
 
 s32 Player_ActionToModelGroup(Player* this, s32 itemAction) {
@@ -1076,6 +1076,7 @@ static s32 sCustomItemActions[] =
     PLAYER_CUSTOM_IA_SPRING_WATER,      /* ITEM_OOT_SPRING_WATER */
     PLAYER_CUSTOM_IA_SPRING_WATER_HOT,  /* ITEM_OOT_SPRING_WATER_HOT */
     PLAYER_CUSTOM_IA_ZORA_EGG,          /* ITEM_OOT_ZORA_EGG */
+    PLAYER_CUSTOM_IA_POWDER_KEG,        /* ITEM_OOT_POWDER_KEG */
 };
 
 s32 Player_CustomItemToItemAction(s32 item, s32 itemAction)
@@ -1127,7 +1128,7 @@ Color_RGB8* Player_ActionToBottleColor(Player* this, s32 itemAction, Color_RGB8*
 s32 Player_CustomActionToBottleExchange(Player* this, s32 itemAction)
 {
     s32 customBottle = itemAction - PLAYER_CUSTOM_IA_MIN;
-    if (customBottle >= 0 && customBottle <= 8)
+    if (customBottle >= 0 && customBottle < 8)
     {
         return customBottle + 6;
     }
@@ -1491,3 +1492,201 @@ void Player_AfterInit(PlayState* play)
 }
 
 PATCH_CALL(0x808452cc, Player_AfterInit);
+
+void Player_CapSpeedXZ(Player* this, f32* speedTarget, s16* yawTarget)
+{
+    s16 yawDiff = this->yaw - *yawTarget;
+
+    f32 step = 1.0f;
+    f32 stepInverse = 0.1f;
+
+    if (Config_Flag(CFG_OOT_AIR_PHYSICS_MM))
+    {
+        if ((this->unk_880 * 1.5f) < fabsf(this->speedXZ)) {
+            step *= 4.0f;
+            stepInverse *= 4.0f;
+        }
+    }
+    else if (this->meleeWeaponState == 0)
+    {
+        this->speedXZ = CLAMP(this->speedXZ, -(R_RUN_SPEED_LIMIT / 100.0f), (R_RUN_SPEED_LIMIT / 100.0f));
+    }
+
+    if (ABS(yawDiff) > 0x6000)
+    {
+        if (Math_StepToF(&this->speedXZ, 0.0f, step))
+        {
+            this->yaw = *yawTarget;
+        }
+    }
+    else
+    {
+        Math_AsymStepToF(&this->speedXZ, *speedTarget, 0.05f, stepInverse);
+        Math_ScaledStepToS(&this->yaw, *yawTarget, 200);
+    }
+}
+
+PATCH_FUNC(0x8083c0a4, Player_CapSpeedXZ)
+
+Actor* Player_SpawnExplosive(ActorContext* actorCtx, Player* player, PlayState* play, s16 actorId, f32 posX, f32 posY, f32 posZ, s16 rotX, s16 rotY, s16 rotZ, s32 params)
+{
+    switch (player->itemAction)
+    {
+    case 0x12: /* PLAYER_IA_BOMB */
+        actorId = ACTOR_EN_BOM; /* ACTOR_CUSTOM_EN_BOM ACTOR_EN_BOM */
+        break;
+    case 0x13: /* PLAYER_IA_BOMBCHU */
+        actorId = ACTOR_EN_BOM_CHU;
+        break;
+    case PLAYER_CUSTOM_IA_POWDER_KEG:
+        actorId = ACTOR_CUSTOM_EN_BOM;
+        if (gCustomSave.powderKegTimer == 0)
+        {
+            gCustomSave.powderKegTimer = 200;
+        }
+        rotX = BOMB_EXPLOSIVE_TYPE_POWDER_KEG;
+        break;
+    }
+    Actor* result = Actor_SpawnAsChild(actorCtx, &player->actor, play, actorId, posX, posY, posZ, rotX, rotY, rotZ, params);
+    if (result == NULL && player->itemAction == PLAYER_CUSTOM_IA_POWDER_KEG)
+    {
+        gCustomSave.powderKegTimer = 0;
+    }
+    return result;
+}
+
+PATCH_CALL(0x808318cc, Player_SpawnExplosive)
+
+s32 Player_ShouldExplosiveError(Player* this, s32 itemAction)
+{
+    s32 ammo;
+    switch (itemAction)
+    {
+    case 0x12: /* PLAYER_IA_BOMB */
+        ammo = gSaveContext.save.info.inventory.ammo[ITS_OOT_BOMBS];
+        break;
+    case 0x13: /* PLAYER_IA_BOMBCHU */
+        ammo = gSaveContext.save.info.inventory.ammo[ITS_OOT_BOMBCHU];
+        break;
+    case PLAYER_CUSTOM_IA_POWDER_KEG:
+        if (Player_GetStrength() < 3)
+            return 1;
+        ammo = gOotExtraAmmo.kegAmmo;
+        break;
+    default:
+        return 0;
+    }
+    return ammo == 0 || gPlay->actorCtx.actors[ACTORCAT_EXPLOSIVE].count >= 3;
+}
+
+/* Original is no longer uses the return value the same way. */
+s32 Player_ActionToExplosive(Player* this, s32 itemAction)
+{
+    switch (itemAction)
+    {
+    case 0x12: /* PLAYER_IA_BOMB */
+    case PLAYER_CUSTOM_IA_POWDER_KEG:
+        return 0;
+    case 0x13: /* PLAYER_IA_BOMBCHU */
+        return 1;
+    default:
+        return -1;
+    }
+}
+
+PATCH_FUNC(0x80079d48, Player_ActionToExplosive)
+
+void Player_DeductExplosiveAmmo(Player* player)
+{
+    switch (player->itemAction)
+    {
+    case 0x12: /* PLAYER_IA_BOMB */
+        Inventory_ChangeAmmo(ITEM_BOMB, -1);
+        break;
+    case 0x13: /* PLAYER_IA_BOMBCHU */
+        Inventory_ChangeAmmo(ITEM_BOMBCHU_10, -1);
+        break;
+    case PLAYER_CUSTOM_IA_POWDER_KEG:
+        DECR(gOotExtraAmmo.kegAmmo);
+        break;
+    }
+}
+
+s32 Player_IsActionCutsceneItem(s8 itemAction)
+{
+    if (itemAction >= 0x1c && itemAction <= 0x1d) /* PLAYER_IA_OCARINA_FAIRY PLAYER_IA_OCARINA_OF_TIME*/
+    {
+        return 1;
+    }
+    if (itemAction >= 0x1f && itemAction <= 0x39) /* PLAYER_IA_BOTTLE_FISH PLAYER_IA_CLAIM_CHECK */
+    {
+        return 1;
+    }
+    if (itemAction >= PLAYER_CUSTOM_IA_MAGIC_MUSHROOM && itemAction <= PLAYER_CUSTOM_IA_ZORA_EGG)
+    {
+        return 1;
+    }
+    return 0;
+}
+
+typedef void (*ItemActionInitFunc)(PlayState*, Player*);
+
+void Player_InvokeItemActionInitFunc(PlayState* play, Player* this, ItemActionInitFunc func)
+{
+    switch (this->itemAction)
+    {
+        case PLAYER_CUSTOM_IA_POWDER_KEG:
+            ItemActionInitFunc Player_InitExplosiveIA = OverlayAddr(0x80831838);
+            Player_InitExplosiveIA(play, this);
+            break;
+        default:
+            func(play, this);
+            break;
+    }
+}
+
+typedef void (*UpperActionFunc)(Player*, PlayState*);
+typedef void (*Player_SetUpperActionFunc)(Player*, UpperActionFunc);
+
+void Player_SetUpperActionFuncToHeldItemAction(Player* this, UpperActionFunc upperActionFunc)
+{
+    Player_SetUpperActionFunc Player_SetUpperActionFunc = OverlayAddr(0x8083164c);
+    switch (this->itemAction)
+    {
+    case PLAYER_CUSTOM_IA_POWDER_KEG:
+        upperActionFunc = OverlayAddr(0x80833770); /* Player_UpperAction_CarryActor */
+        break;
+    default:
+        break;
+    }
+    Player_SetUpperActionFunc(this, upperActionFunc);
+}
+
+PATCH_CALL(0x808326e4, Player_SetUpperActionFuncToHeldItemAction)
+PATCH_CALL(0x80832b28, Player_SetUpperActionFuncToHeldItemAction)
+PATCH_CALL(0x80832d3c, Player_SetUpperActionFuncToHeldItemAction)
+
+s32 Player_ShouldDamageRecoil(Player* this, s32* shieldHit)
+{
+    *shieldHit = 0;
+    if (this->shieldQuad.base.acFlags & AC_BOUNCED)
+    {
+        *shieldHit = 1;
+        return 1;
+    }
+
+    if (this->invincibilityTimer < 0 && this->cylinder.base.acFlags & AC_HIT)
+    {
+        if (this->cylinder.elem.atHit != NULL && this->cylinder.elem.atHit->atFlags & 0x20000000)
+        {
+            return 1;
+        }
+        if (this->cylinder.base.ac != NULL && this->cylinder.base.ac->id == ACTOR_CUSTOM_EN_BOM
+            && this->cylinder.elem.acHitElem != NULL && this->cylinder.elem.acHitElem->atDmgInfo.dmgFlags != DMG_UNBLOCKABLE)
+        {
+            return 1;
+        }
+    }
+
+    return 0;
+}